test(vdb): consolidate bounded sink contracts
